+++ Cult of the Lamb/SkipOfTheLamb/GlobalUsings.cs
@@ -1,8 +1,21 @@
 // Global using directives
 
+global using System;
+global using System.Collections;
+global using System.Collections.Generic;
+global using System.Diagnostics;
+global using System.IO;
+global using System.Linq;
 global using System.Reflection;
+global using System.Reflection.Emit;
 global using BepInEx;
+global using BepInEx.Bootstrap;
 global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using HarmonyLib;
+global using MMBiomeGeneration;
 global using MMTools;
+global using MonoMod.Utils;
+global using Shared;
+global using UnityEngine;
+global using Logger = UnityEngine.Logger;